Combines both unit test files to run in coverage report, adds .coveragerc
