+++ Tests/.coveragerc
@@ -1,0 +1,7 @@
+# .coveragerc to control coverage.py
+[run]
+omit =
+    */site-packages/*
+    */distutils/*
+    */Lib/*
+    Tests/*
